Done changes related to move elasticsearch configuration in json
